data_utils.py restructured as multiple files in the /data folder
