+++ tests/conftest.py
@@ -9,7 +9,7 @@ from flask_script import Manager
 
 from application import db as app_db
 from application.auth.models import *
-from application.cms.data_utils import Harmoniser
+from application.data.standardisers.value_category_standardiser import Harmoniser
 from application.cms.models import *
 from application.config import TestConfig
 from application.factory import create_app
